+++ src/lib/containers/index.tsx
@@ -1,9 +1,11 @@
 import Artist from "./artist"
 import Gene from "./gene"
 import Home from "./home"
+import WorksForYou from "./works_for_you"
 
 export default {
   Artist,
   Home,
   Gene,
+  WorksForYou,
 }